Finish bottom bar and apps page
