+++ NeuroAccessMaui/Resources/Languages/AppResources.es.resx
@@ -2822,4 +2822,16 @@
   <data name="BetaFeaturesDescription" xml:space="preserve">
     <value>Las funciones beta están en desarrollo y podrían carecer de funcionalidad o diseño. Sin embargo, algunas de sus funciones principales ya están disponibles.</value>
   </data>
+  <data name="SmartThings" xml:space="preserve">
+    <value>Cosas inteligentes</value>
+  </data>
+  <data name="Assets" xml:space="preserve">
+    <value>Activos</value>
+  </data>
+  <data name="MicroLoan" xml:space="preserve">
+    <value>Micropréstamo</value>
+  </data>
+  <data name="NFC" xml:space="preserve">
+    <value>NFC</value>
+  </data>
 </root>
