+++ NeuroAccessMaui/Resources/Languages/AppResources.pt.resx
@@ -2822,4 +2822,16 @@
   <data name="BetaFeaturesDescription" xml:space="preserve">
     <value>Os recursos beta estão em desenvolvimento e podem não ter funcionalidade ou design completos. No entanto, alguns de seus recursos principais já estão disponíveis.</value>
   </data>
+  <data name="SmartThings" xml:space="preserve">
+    <value>Coisas inteligentes</value>
+  </data>
+  <data name="Assets" xml:space="preserve">
+    <value>Ativos</value>
+  </data>
+  <data name="MicroLoan" xml:space="preserve">
+    <value>Microempréstimo</value>
+  </data>
+  <data name="NFC" xml:space="preserve">
+    <value>NFC</value>
+  </data>
 </root>
