+++ NeuroAccessMaui/Resources/Languages/AppResources.resx
@@ -2822,4 +2822,16 @@
   <data name="BetaFeaturesDescription" xml:space="preserve">
     <value>Beta features are under development and could lack functionality or design. However, some of their core features are available</value>
   </data>
+  <data name="SmartThings" xml:space="preserve">
+    <value>Smart Things</value>
+  </data>
+  <data name="Assets" xml:space="preserve">
+    <value>Assets</value>
+  </data>
+  <data name="MicroLoan" xml:space="preserve">
+    <value>MicroLoan</value>
+  </data>
+  <data name="NFC" xml:space="preserve">
+    <value>NFC</value>
+  </data>
 </root>
