+++ NeuroAccessMaui/UI/Pages/Main/Apps/AppsViewModel.cs
@@ -1,7 +1,16 @@
 ﻿
 using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
 using NeuroAccessMaui.Services;
+using NeuroAccessMaui.Services.UI;
 using NeuroAccessMaui.Services.Tag;
+using NeuroAccessMaui.UI.Pages.Contacts.MyContacts;
+using NeuroAccessMaui.UI.Pages.Contracts.MyContracts;
+using NeuroAccessMaui.UI.Pages.Applications.Applications;
+using NeuroAccessMaui.UI.Pages.Things.MyThings;
+using NeuroAccessMaui.UI.Pages.Identity.ViewIdentity;
+using EDaler;
+using NeuroAccessMaui.UI.Pages.Wallet.MyWallet;
 
 namespace NeuroAccessMaui.UI.Pages.Main.Apps
 {
@@ -20,9 +29,143 @@ namespace NeuroAccessMaui.UI.Pages.Main.Apps
 			set => this.hasBetaFeatures = value;
 		}
 
-
 		// Binding for selecting between NeuroIconButton and NeuroIconButtonDisabled style depending on if has beta features enabled
 		public Style BetaButtonStyle => this.HasBetaFeatures ? AppStyles.NeuroIconButton : AppStyles.NeuroIconButtonDisabled;
 
+		#region Navigation Commands
+
+		[RelayCommand]
+		private static async Task ShowContacts()
+		{
+			try
+			{
+				ContactListNavigationArgs Args = new();
+				await ServiceRef.UiService.GoToAsync(nameof(MyContactsPage), Args, BackMethod.Pop);
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
+		}
+
+		[RelayCommand]
+		private static async Task ShowMyContracts()
+		{
+			try
+			{
+				MyContractsNavigationArgs Args = new(ContractsListMode.Contracts);
+				await ServiceRef.UiService.GoToAsync(nameof(MyContractsPage), Args, BackMethod.Pop);
+			}
+			catch (Exception ex)
+			{
+				ServiceRef.LogService.LogException(ex);
+			}
+		}
+
+		[RelayCommand]
+		private static async Task ShowNewContract()
+		{
+			try
+			{
+				MyContractsNavigationArgs Args = new(ContractsListMode.ContractTemplates);
+				await ServiceRef.UiService.GoToAsync(nameof(MyContractsPage), Args, BackMethod.Pop);
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
+		}
+
+		[RelayCommand]
+		private static async Task ShowNewToken()
+		{
+			try
+			{
+				MyContractsNavigationArgs Args = new(ContractsListMode.TokenCreationTemplates);
+				await ServiceRef.UiService.GoToAsync(nameof(MyContractsPage), Args, BackMethod.Pop);
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
+		}
+
+		[RelayCommand]
+		private static async Task ShowApplications()
+		{
+			try
+			{
+				await ServiceRef.UiService.GoToAsync(nameof(ApplicationsPage));
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
+		}
+
+		[RelayCommand]
+		private static async Task ShowThings()
+		{
+			try
+			{
+				await ServiceRef.UiService.GoToAsync(nameof(MyThingsPage));
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
+		}
+
+		[RelayCommand]
+		private static async Task ShowId()
+		{
+			try
+			{
+				if (await App.AuthenticateUserAsync(AuthenticationPurpose.ViewId))
+					await ServiceRef.UiService.GoToAsync(nameof(ViewIdentityPage));
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
+		}
+
+		[RelayCommand]
+		internal static async Task ShowWallet()
+		{
+			try
+			{
+				Balance Balance = await ServiceRef.XmppService.GetEDalerBalance();
+				(decimal PendingAmount, string PendingCurrency, PendingPayment[] PendingPayments) = await ServiceRef.XmppService.GetPendingEDalerPayments();
+				(AccountEvent[] Events, bool More) = await ServiceRef.XmppService.GetEDalerAccountEvents(Constants.BatchSizes.AccountEventBatchSize);
+
+				WalletNavigationArgs Args = new(Balance, PendingAmount, PendingCurrency, PendingPayments, Events, More);
+
+				await ServiceRef.UiService.GoToAsync(nameof(MyEDalerWalletPage), Args, BackMethod.Pop);
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+				await ServiceRef.UiService.DisplayException(Ex);
+			}
+		}
+
+		[RelayCommand]
+        public async Task ViewMainPage()
+        {
+            try
+            {
+                if (Application.Current?.MainPage?.Navigation != null)
+                {
+                    await Application.Current.MainPage.Navigation.PopToRootAsync();
+                }
+            }
+            catch (Exception Ex)
+            {
+                ServiceRef.LogService.LogException(Ex);
+            }
+        }
+    
+		#endregion
 	}
 }
