+++ NeuroAccessMaui/UI/Pages/Main/MainViewModel.cs
@@ -16,6 +16,10 @@ using NeuroAccessMaui.Services.Theme;
 using Waher.Runtime.Inventory;
 using System.Reflection.Metadata;
 using NeuroAccessMaui.UI.Pages.Main.Apps;
+using EDaler;
+using NeuroAccessMaui.UI.Pages.Wallet.MyWallet;
+using NeuroAccessMaui.Services.UI;
+using NeuroAccessMaui.UI.Pages.Main.Settings;
 
 namespace NeuroAccessMaui.UI.Pages.Main
 {
@@ -180,13 +184,40 @@ namespace NeuroAccessMaui.UI.Pages.Main
 			}
 		}
 
-		[RelayCommand(AllowConcurrentExecutions = false)]
-		public static async Task OpenFlyout()
+		[RelayCommand]
+		internal static async Task ShowWallet()
 		{
-			await MainThread.InvokeOnMainThreadAsync(() =>
-				Shell.Current.FlyoutIsPresented = true
-			);
+			try
+			{
+				Balance Balance = await ServiceRef.XmppService.GetEDalerBalance();
+				(decimal PendingAmount, string PendingCurrency, PendingPayment[] PendingPayments) = await ServiceRef.XmppService.GetPendingEDalerPayments();
+				(AccountEvent[] Events, bool More) = await ServiceRef.XmppService.GetEDalerAccountEvents(Constants.BatchSizes.AccountEventBatchSize);
 
+				WalletNavigationArgs Args = new(Balance, PendingAmount, PendingCurrency, PendingPayments, Events, More);
+
+				await ServiceRef.UiService.GoToAsync(nameof(MyEDalerWalletPage), Args, BackMethod.Pop);
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+				await ServiceRef.UiService.DisplayException(Ex);
+			}
+		}
+
+		/// <summary>
+		/// Shows the settings page.
+		/// </summary>
+		[RelayCommand]
+		private static async Task ShowSettings()
+		{
+			try
+			{
+				await ServiceRef.UiService.GoToAsync(nameof(SettingsPage));
+			}
+			catch (Exception Ex)
+			{
+				ServiceRef.LogService.LogException(Ex);
+			}
 		}
 	}
 }
